requirements.txt
Flask~=3.0.0
Flask-PyMongo
pymongo~=4.6.0
python-dotenv~=1.0.0
requests~=2.31.0
beautifulsoup4~=4.12.2
PyJWT~=2.8.0
Werkzeug~=3.0.1
mongomock~=4.1.2
pytest~=7.4.3